backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0
metpy>=1.5.0
herbie-data>=2024.3.0
xarray>=2023.0.0
